changed the osmids to be ints instead of long to save memory, also changed the logic of creating the osmids in Green Paths 2.0 code, to not to produce too large/long integers for native int

+++ src/main/java/com/conveyal/r5/analyst/TravelTimeComputer.java
@@ -244,7 +244,7 @@ public class TravelTimeComputer {
             {
                 // only get osmids if the streetlayer has cost fields
                 if (network.streetLayer.edgeStore.costFields != null) {
-                    List<List<Long>> osmIdResults = CustomCost.getOsmIdsFromRouterState(nonTransitTravelTimesToDestinations, sr, network);
+                    List<List<Integer>> osmIdResults = CustomCost.getOsmIdsFromRouterState(nonTransitTravelTimesToDestinations, sr, network);
                     travelTimeReducer.setOsmIdsResult(osmIdResults);
                 }
             }
